Code improvement

+++ plugins/robots/common/twoDModel/src/engine/view/parts/colorItemPopup.h
@@ -47,7 +47,7 @@ public:
 	QColor lastColor() const;
 
 	/// Returns last pen thickness entered by user.
-	int lastThickness() const;
+	qreal lastThickness() const;
 
 Q_SIGNALS:
 	/// Emitted when pen configured by user changed.
